# apps/web/lib/contracts.ts
import type { components } from "./api-schema";

export type ModelInfo = components["schemas"]["ModelInfo"];
export type PendingIntervention =
  components["schemas"]["PendingIntervention"];
export type RunDetail = components["schemas"]["RunDetail"];
export type RunEvent = components["schemas"]["RunEvent"];
export type ScenarioManifest = components["schemas"]["ScenarioManifest"];
export type ScreenshotArtifact =
  components["schemas"]["ScreenshotArtifact"];
export type StartRunRequest = components["schemas"]["StartRunRequest"];
export type StartRunResponse = components["schemas"]["StartRunResponse"];